Re-organize project to export package
