+++ planet-clicker/Assets/LibplanetUnity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3d3a48c4354a1424e9d67ac98a4c0316
+guid: 337e7b5ffc6f8c24d8d3baa6561914b9
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
