+++ planet-clicker/Assets/LibplanetUnity/Action.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4a7a7276efd23450e95e223cd4967437
+guid: 2b232be6b73071940b27f820b65fe066
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
