+++ planet-clicker/Assets/LibplanetUnity/Editor/LibplanetEditor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 18959332d09ca6a48b925018da69a840
+guid: 16e7ef3dfd5727544903120b45c8328a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
