+++ planet-clicker/Assets/LibplanetUnity/Helper.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a0898c7f2fca849c898e4ca840dcfc5d
+guid: 0c73297c46de7e14fa2ac3f00fe24046
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
