+++ planet-clicker/Assets/LibplanetUnity/MonoSingleton.cs
@@ -1,14 +1,7 @@
 using UnityEngine;
 
-namespace _Script
+namespace LibplanetUnity
 {
-    /// <summary>
-    /// ToDo. 루트 게임오브젝트가 아닐 경우 `DontDestroyOnLoad(gameObject);`가 작동하지 않는다.
-    /// 특정 게임오브젝트의 자식으로 존재하는 싱글턴 게임오브젝트가 필요할까?
-    /// 부모 게임오브젝트가 파괴되었을 때 싱글턴 게임오브젝트는 어떤 형태로 존재해야 하는가?
-    /// 필요하다면, 어떻게 제공할까?
-    /// </summary>
-    /// <typeparam name="T"></typeparam>
     public class MonoSingleton<T> : MonoBehaviour where T : MonoSingleton<T>
     {
         private static T _instance;
@@ -78,14 +71,6 @@ namespace _Script
             }
         }
 
-        /// <summary>
-        /// When Unity quits, it destroys objects in a random order.
-        /// In principle, a Singleton is only destroyed when application quits.
-        /// If any script calls Instance after it have been destroyed, 
-        ///   it will create a buggy ghost object that will stay on the Editor scene
-        ///   even after stopping playing the Application. Really bad!
-        /// So, this was made to be sure we're not creating that buggy ghost object.
-        /// </summary>
         protected virtual void OnDestroy()
         {
         }
@@ -97,7 +82,6 @@ namespace _Script
 
         #endregion
 
-        // instance를 생성하기 위한 빈 메소드.
         public void EmptyMethod()
         {
