+++ planet-clicker/Assets/LibplanetUnity/MonoSingleton.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 18959332d09ca6a48b925018da69a840
+guid: f461157903064214a88cc8422b38bcc4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
