+++ planet-clicker/Assets/_Script/Action.meta
@@ -1,3 +1,8 @@
-﻿fileFormatVersion: 2
+fileFormatVersion: 2
 guid: 62bdb5a4bbe14b52b6c383871a31de69
-timeCreated: 1568945736
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
