+++ planet-clicker/Assets/_Script/Action/AddCount.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 6989f23aebd54703a26892f20013c034
-timeCreated: 1568945976
+fileFormatVersion: 2
+guid: 33cd1de38848c9f45a854226cd886488
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
