--- planet-clicker/Assets/_Script/BlockChain/AgentController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 78ff9ea9564e41dd9d8963a2be208feb
-timeCreated: 1568945792
