--- planet-clicker/Assets/_Script/Helper/Options.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1cb988400d174a7ea44afce6019e4f16
-timeCreated: 1568946456
