--- planet-clicker/Assets/_Script/MonoSingleton.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 65ec507c7cb34328a2cf5093ba140901
-timeCreated: 1568945851
